+++ packages/liveblocks-react-comments/src/components/InboxNotification.tsx
@@ -18,12 +18,16 @@ import type {
   ComponentProps,
   ComponentPropsWithoutRef,
   ComponentType,
+  MouseEvent,
   ReactNode,
+  SyntheticEvent,
 } from "react";
-import React, { forwardRef, useMemo } from "react";
+import React, { forwardRef, useCallback, useMemo, useState } from "react";
 
 import type { GlobalComponents } from "../components";
 import { useComponents } from "../components";
+import { CheckIcon } from "../icons/Check";
+import { EllipsisIcon } from "../icons/Ellipsis";
 import type {
   CommentOverrides,
   GlobalOverrides,
@@ -40,8 +44,11 @@ import {
   CommentNonInteractiveReaction,
 } from "./Comment";
 import { Avatar, type AvatarProps } from "./internal/Avatar";
+import { Button } from "./internal/Button";
+import { Dropdown, DropdownItem, DropdownTrigger } from "./internal/Dropdown";
 import { List } from "./internal/List";
 import { Room } from "./internal/Room";
+import { Tooltip } from "./internal/Tooltip";
 import { User } from "./internal/User";
 
 const INBOX_NOTIFICATION_THREAD_MAX_COMMENTS = 3;
@@ -81,7 +88,16 @@ type InboxNotificationKindsWithRef = AddRefToComponents<
   ComponentProps<"a">["ref"]
 >;
 
-export interface InboxNotificationProps extends ComponentPropsWithoutRef<"a"> {
+interface InboxNotificationSharedProps {
+  /**
+   * How to show or hide the actions.
+   */
+  showActions?: boolean | "hover";
+}
+
+export interface InboxNotificationProps
+  extends ComponentPropsWithoutRef<"a">,
+    InboxNotificationSharedProps {
   /**
    * The inbox notification to display.
    */
@@ -106,7 +122,8 @@ export interface InboxNotificationProps extends ComponentPropsWithoutRef<"a"> {
 }
 
 export interface InboxNotificationThreadProps
-  extends Omit<InboxNotificationProps, "kinds"> {
+  extends Omit<InboxNotificationProps, "kinds">,
+    InboxNotificationSharedProps {
   /**
    * Whether to show the room name in the title.
    */
@@ -114,7 +131,9 @@ export interface InboxNotificationThreadProps
 }
 
 interface InboxNotificationLayoutProps
-  extends Omit<ComponentPropsWithoutRef<"a">, "title"> {
+  extends Omit<ComponentPropsWithoutRef<"a">, "title">,
+    InboxNotificationSharedProps {
+  inboxNotificationId: string;
   aside: ReactNode;
   title: ReactNode;
   date: Date | string | number;
@@ -137,11 +156,13 @@ const InboxNotificationLayout = forwardRef<
 >(
   (
     {
+      inboxNotificationId,
       children,
       aside,
       title,
       date,
       unread,
+      showActions,
       overrides,
       components,
       className,
@@ -151,11 +172,42 @@ const InboxNotificationLayout = forwardRef<
   ) => {
     const $ = useOverrides(overrides);
     const { Anchor } = useComponents(components);
+    const [isMoreActionOpen, setMoreActionOpen] = useState(false);
+    const { useMarkInboxNotificationAsRead } = useLiveblocksContextBundle();
+    const markInboxNotificationAsRead = useMarkInboxNotificationAsRead();
+
+    const stopPropagation = useCallback((event: SyntheticEvent) => {
+      event.stopPropagation();
+    }, []);
+
+    const preventDefaultAndStopPropagation = useCallback(
+      (event: SyntheticEvent) => {
+        event.preventDefault();
+        event.stopPropagation();
+      },
+      []
+    );
+
+    const handleMoreClick = useCallback((event: MouseEvent) => {
+      event.preventDefault();
+      event.stopPropagation();
+      setMoreActionOpen((open) => !open);
+    }, []);
+
+    const handleMarkAsRead = useCallback(() => {
+      markInboxNotificationAsRead(inboxNotificationId);
+    }, [inboxNotificationId, markInboxNotificationAsRead]);
 
     return (
       <TooltipProvider>
         <Anchor
-          className={classNames("lb-root lb-inbox-notification", className)}
+          className={classNames(
+            "lb-root lb-inbox-notification",
+            showActions === "hover" &&
+              "lb-inbox-notification:show-actions-hover",
+            isMoreActionOpen && "lb-inbox-notification:action-open",
+            className
+          )}
           dir={$.dir}
           data-unread={unread ? "" : undefined}
           {...props}
@@ -180,6 +232,41 @@ const InboxNotificationLayout = forwardRef<
                   )}
                 </span>
               </div>
+              {showActions && (
+                <div className="lb-inbox-notification-actions">
+                  <Dropdown
+                    open={isMoreActionOpen}
+                    onOpenChange={setMoreActionOpen}
+                    align="end"
+                    content={
+                      <>
+                        <DropdownItem
+                          onSelect={handleMarkAsRead}
+                          onClick={stopPropagation}
+                          disabled={!unread}
+                        >
+                          <CheckIcon className="lb-dropdown-item-icon" />
+                          {$.INBOX_NOTIFICATION_MARK_AS_READ}
+                        </DropdownItem>
+                      </>
+                    }
+                  >
+                    <Tooltip content={$.INBOX_NOTIFICATION_MORE}>
+                      <DropdownTrigger asChild>
+                        <Button
+                          className="lb-inbox-notification-action"
+                          onClick={handleMoreClick}
+                          onPointerDown={preventDefaultAndStopPropagation}
+                          onPointerUp={preventDefaultAndStopPropagation}
+                          aria-label={$.INBOX_NOTIFICATION_MORE}
+                        >
+                          <EllipsisIcon className="lb-button-icon" />
+                        </Button>
+                      </DropdownTrigger>
+                    </Tooltip>
+                  </Dropdown>
+                </div>
+              )}
             </div>
             <div className="lb-inbox-notification-body">{children}</div>
           </div>
@@ -357,7 +444,14 @@ const InboxNotificationThread = forwardRef<
   InboxNotificationThreadProps
 >(
   (
-    { inboxNotification, href, showRoomName = true, overrides, ...props },
+    {
+      inboxNotification,
+      href,
+      showRoomName = true,
+      showActions = "hover",
+      overrides,
+      ...props
+    },
     forwardedRef
   ) => {
     const $ = useOverrides(overrides);
@@ -474,12 +568,14 @@ const InboxNotificationThread = forwardRef<
 
     return (
       <InboxNotificationLayout
+        inboxNotificationId={inboxNotification.id}
         aside={aside}
         title={title}
         date={date}
         unread={unread}
         overrides={overrides}
         href={resolvedHref}
+        showActions={showActions}
         {...props}
         ref={forwardedRef}
       >