+++ CHANGELOG.md
@@ -2,15 +2,23 @@
 
 ### `@liveblocks/react`
 
-### `@liveblocks/node`
-
-- Add `updateRoomId` method that lets you update room id of the specified room.
 - Add `scrollOnLoad` option to `useThreads`: enabled by default, this option
   controls whether to scroll to a comment on load based on the URL hash.
 - `useUser` and `useRoomInfo` no longer support returning nothing. Returning
   `undefined` will now be treated as an error.
 - Fix bug where `useUser` and `useRoomInfo` returned an extra `data` superfluous
   property.
+
+### `@liveblocks/react-comments`
+
+- Add actions to `InboxNotification` with a single action for now: marking as
+  read.
+- Improve actions hover behavior in `Comment`/`Thread`.
+
+### `@liveblocks/node`
+
+- Add `updateRoomId` method that lets you update the room ID of the specified
+  room.
 
 # v1.10.3